+++ scratch.js
@@ -1,8 +1,7 @@
 'use strict';
 var nlp = require('./src/index');
 // nlp.verbose('tagger');
-// const corpus = require('nlp-corpus');
-// let text = corpus.sotu.parsed()[0];
 // const fresh = require('./test/unit/lib/freshPrince.js');
 
-let r = nlp('buy eggs constructor prototype');
+var r = nlp('they are flowers');
+r.debug();