+++ test/unit/subset/value/value.test.js
@@ -36,6 +36,7 @@ test('value-to_ordinal:', function (t) {
   t.end();
 });
 
+
 test('value-number:', function (t) {
   [
     ['five hundred feet', 500],
@@ -52,28 +53,84 @@ test('value-number:', function (t) {
   t.end();
 });
 
-// test('value-unit:', function(t) {
-//   [
-//     ['five hundred feet', 'feet'],
-//     ['fifty hertz', 'hertz'],
-//     ['100 dollars', 'dollars'],
-//     ['$100', 'dollar'],
-//     ['¥2.5', 'yen'],
-//     ['€3,000,100', 'euro'],
-//     ['EUR 9.99', 'eur'],
-//     ['5 g', 'g'],
-//     ['3 grams', 'gram'],
-//     ['2 inches', 'inch'],
-//     ['2 in', 'in'],
-//     ['5 g sugar', 'g'],
-//     ['10 grams of sugar', 'gram'],
-//     ['fifty inches of snow', 'inch'],
-//   ].forEach(function (a) {
-//     var str = nlp.value(a[0]).unit;
-//     str_test(str, a[0], a[1], t);
-//   });
-//   t.end();
-// });
+test('add/subtract:', function (t) {
+  var r = nlp('beginning of 2019').values().add(2).all();
+  t.equal(r.out(), 'beginning of 2021', 'add-2-cardinal');
+
+  r = nlp('beginning of the 2019th').values().add(2).all();
+  t.equal(r.out(), 'beginning of the 2021st', 'add-2-ordinal');
+
+  r = nlp('beginning of the 2019th').values().add(-2).all();
+  t.equal(r.out(), 'beginning of the 2017th', 'add-minus-2-ordinal');
+
+  r = nlp('beginning of 2019').values().subtract(2).all();
+  t.equal(r.out(), 'beginning of 2017', 'subtract-2-cardinal');
+
+  r = nlp('beginning of the 2019th').values().subtract(2).all();
+  t.equal(r.out(), 'beginning of the 2017th', 'subtract-2-ordinal');
+
+  r = nlp('seventeen years old').values().add(2).all();
+  t.equal(r.out(), 'nineteen years old', 'text-add-2-ordinal');
+  r = nlp('seventeenth birthday').values().add(2).all();
+  t.equal(r.out(), 'nineteenth birthday', 'text-add-2-ordinal');
+
+  r = nlp('seventeen years old').values().subtract(2).all();
+  t.equal(r.out(), 'fifteen years old', 'text-subtract-2-cardinal');
+  r = nlp('seventeenth birthday').values().subtract(2).all();
+  t.equal(r.out(), 'fifteenth birthday', 'text-subtract-2-cardinal');
+
+  r = nlp('seven apples and 1,231 peaches').values().add(50).all();
+  t.equal(r.out(), 'fifty seven apples and 1,281 peaches', 'two-add-50s');
+  t.end();
+});
+
+test('increment:', function (t) {
+  var r = nlp('seven apples and 231 peaches');
+  r.values().increment();
+  t.equal(r.out(), 'eight apples and 232 peaches', 'increment-cardinal');
+  r.values().decrement();
+  t.equal(r.out(), 'seven apples and 231 peaches', 'decrement-cardinal');
+
+  r = nlp('seventh place and 12th place');
+  r.values().increment().increment();
+  t.equal(r.out(), 'ninth place and 14th place', 'increment-ordinal');
+  r.values().decrement().decrement();
+  t.equal(r.out(), 'seventh place and 12th place', 'decrement-ordinal');
+  t.end();
+});
+
+test('nounit:', function (t) {
+  var r = nlp('seven apples and 231 peaches');
+  var arr = r.values().out('array');
+  t.deepEqual(arr, ['seven apples', '231 peaches']);
+
+  arr = r.values().noUnits().out('array');
+  t.deepEqual(arr, ['seven', '231']);
+  t.end();
+});
+
+test('value-unit:', function(t) {
+  [
+    ['five hundred feet', 'feet'],
+    // ['fifty hertz', 'hertz'],
+    ['100 dollars', 'dollars'],
+    // ['$100', 'dollar'],
+    // ['¥2.5', 'yen'],
+    // ['€3,000,100', 'euro'],
+    // ['EUR 9.99', 'eur'],
+    // ['5 g', 'g'],
+    // ['2 in', 'in'],
+    // ['5 g sugar', 'g'],
+    ['3 grams', 'grams'],
+    ['2 inches', 'inches'],
+    ['10 grams of sugar', 'grams'],
+    ['fifty inches of snow', 'inches'],
+  ].forEach(function (a) {
+    var r = nlp(a[0]).values().units();
+    str_test(r.out('normal'), a[0], a[1], t);
+  });
+  t.end();
+});
 //
 // test('value-measurement:', function(t) {
 //   [